rename awss3 module to reflect provider of SpNetwork data
