+++ flask_app/analyst/count.py
@@ -6,7 +6,7 @@ from flask_app.common.s2n_type import APIService, AnalystOutput
 from flask_app.analyst.base import _AnalystService
 
 from sppy.aws.aws_constants import PROJ_BUCKET
-from sppy.tools.provider.awss3 import SpNetAnalyses
+from sppy.tools.provider.spnet import SpNetAnalyses
 from sppy.tools.s2n.utils import (combine_errinfo, get_traceback)
